rt-error: added conditional to info() and debug()
This change was motivated by the desire to clean up the output of
cyclicdeadline, and include the extra info only when requested.

Instead of having to check if the program is in debugging mode, the
rt-error functions will automatically check by passing in an argument.

The other changes in this patch edit the function calls to debug() and
info() to work with the changes in rt-error.

Signed-off-by: Anubhav Shelat <ashelat@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>