rt-tests: hackbench: removed extra use of optind
Currently, using the -s option displays the usage message, even if the
option is properly used.

This is because Commit 778a02b7c519 ("rt-tests: hackbench: drop incorrect
and unnecessary usage of optind") forgot to drop a use of optind when
processing option 's' which was fixed in this commit.

Now the -s option works correctly with the proper arguments.

Note: The next commit in this patchset fixes "ERROR: do not use
assignment in if condition" on line 459.

Signed-off-by: Anubhav Shelat <ashelat@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>