cyclictest.8: Standardize --verbose paragraph formatting
All of cyclictest's options are described in the manpage through a single
paragraph except --verbose, which renders the verbose statistics output
format in a separate paragraph for no special reason.

Let the description text of --versbose match the other parameters style.
Beside consistency, this will also be useful when all the options will be
separated at the manpage source level.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>
Signed-off-by: John Kacur <jkacur@redhat.com>