cyclictest.8: Rework the Synopsis blurb
The cyclictest's SYNOPSIS blurb is buggy and incomplete.

For the cyclictest's flag list:

  - Remove 'f', for --fifo, as it's not a flag and requires the fifo's path
  - Add 'N', for --nsecs
  - Add 'R', for --resolution
  - Add 'S', for --smp
  - Add 'u', for --unbuffered
  - Add 'x', for --posix_timers

For options with arguments, expand the argument names with something
comprehensible since the whole purpose of the SYNOPSIS section is to get,
in a blitz, an understanding of the tool's actual options.  The original
argument names, especially when -combined with the one-letter option's
shorthand form, were totally -incomprehensible.

For consistency, be coherent regarding the arguments' names; e.g.,
"*-path" for all file paths like file-path and fifo-path, "proc-set" for
cpuset-related arguments, "num-*" for counts like num-threads and
num-loops; "*-usec" for microsecond arguments, and so on.

Also add the missing options --deepest-idle-state, --histfile, --latency,
--mainaffinity, --policy, --spike, --spike-nodes, --dbg_cyclictest,
--default-system, --smi, --priospread, --smi, and --tracemark.  Do not
add --laptop since it is deprecated.

Note, this makes the cyclictest's SYNOPSIS blurb complete and gives a
comprehensive view of cyclictest's abilities in a glance.

Signed-off-by: Ahmed S. Darwish <darwi@linutronix.de>
Signed-off-by: John Kacur <jkacur@redhat.com>