rt-utils: Rename get_debugfileprefix() and let it look for tracefs
get_debugfileprefix() used to look for debugfs and added the tracing
suffix to it. In the early days it was not always mounted at the same
destination. Later, that function got a fastpath to look directly at
"/sys/kernel/debug/tracing". This avoids parsing the mounted filesystem
and looking for debugfs.

Since about v4.1 (released over 10 years ago) that tracing folder moved
out debugfs and got its own filesystem (tracefs) and is mounted at
/sys/kernel/tracing. Since then that /sys/kernel/debug/tracing folder is
just an automount/ alias to /sys/kernel/tracing.
Since v6.17 that automount can be disabled and is schedulued for
removal.
Strart using tracefs directly instead that alias:
- Replace that "debug" prefix with tracefs since debugfs is no longer
  the thing that is used
- Look directy for /sys/kernel/tracing
- If it is not there, look for tracefs filesystem in case it is mounted
  at an unusual spot.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
- a few quick spelling and grammar fixes
Signed-off-by: John Kacur <jkacur@redhat.com>