rt-utils: Decouple writing trace-marks with trace stopping
Most users write a trace mark and expect that the trace stops afterwards.
Decoupling writing and stopping the trace so it can be used in context
where trace stopping is not needed.

Make tracemark() writing just to the tracemark file.
Add tracing_stop() which stops tracing and let current users of
tracemark() use it.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
- Fixed spelling of "writing" and "afterwards"
Signed-off-by: John Kacur <jkacur@redhat.com>